Added authentication to terms and lessons
Deleted junk files, added html to terms, and made vague attempt to fix tests
