+++ app/views/admin/terms/_form.html.erb
@@ -10,13 +10,11 @@
       </ul>
     </div>
   <% end %>
-
-  <div class="field">
-    <%= f.label :name %>
-    <%= f.text_field :name %>
-  </div>
-
-  <div class="actions">
-    <%= f.submit %>
-  </div>
+  <dl id="meta-top">
+    <dt><label for="term-name">Name</label></dt>
+      <dd>
+        <%= f.text_field :name %>
+      </dd>
+  </dl>
+  <%= f.submit %>
 <% end %>
