+++ spec/examples.txt
@@ -1,0 +1,36 @@
+example_id                                                   | status | run_time        |
+------------------------------------------------------------ | ------ | --------------- |
+./spec/controllers/admin/lessons_controller_spec.rb[1:1:1]   | failed | 0.0308 seconds  |
+./spec/controllers/admin/lessons_controller_spec.rb[1:1:2]   | failed | 0.03002 seconds |
+./spec/controllers/admin/lessons_controller_spec.rb[1:1:3]   | failed | 0.03249 seconds |
+./spec/controllers/admin/lessons_controller_spec.rb[1:2:1]   | failed | 0.33706 seconds |
+./spec/controllers/admin/lessons_controller_spec.rb[1:2:2]   | failed | 0.0423 seconds  |
+./spec/controllers/admin/lessons_controller_spec.rb[1:2:3]   | failed | 0.03376 seconds |
+./spec/controllers/admin/lessons_controller_spec.rb[1:2:4]   | failed | 0.04667 seconds |
+./spec/controllers/admin/lessons_controller_spec.rb[1:2:5]   | failed | 0.04239 seconds |
+./spec/controllers/admin/lessons_controller_spec.rb[1:2:6]   | failed | 0.05133 seconds |
+./spec/controllers/admin/terms_controller_spec.rb[1:1:1:1]   | failed | 0.02437 seconds |
+./spec/controllers/admin/terms_controller_spec.rb[1:1:2:1]   | failed | 0.02443 seconds |
+./spec/controllers/admin/terms_controller_spec.rb[1:1:3:1]   | failed | 0.0197 seconds  |
+./spec/controllers/admin/terms_controller_spec.rb[1:2:1:1]   | failed | 0.03088 seconds |
+./spec/controllers/admin/terms_controller_spec.rb[1:2:2:1:1] | failed | 0.03179 seconds |
+./spec/controllers/admin/terms_controller_spec.rb[1:2:2:1:2] | failed | 0.10306 seconds |
+./spec/controllers/admin/terms_controller_spec.rb[1:2:2:2:1] | failed | 0.02033 seconds |
+./spec/controllers/admin/terms_controller_spec.rb[1:2:3:1:1] | failed | 0.02646 seconds |
+./spec/controllers/admin/terms_controller_spec.rb[1:2:3:1:2] | failed | 0.02465 seconds |
+./spec/controllers/admin/terms_controller_spec.rb[1:2:3:2:1] | failed | 0.03135 seconds |
+./spec/controllers/admin/terms_controller_spec.rb[1:2:4:1]   | failed | 0.02417 seconds |
+./spec/controllers/admin/terms_controller_spec.rb[1:2:4:2]   | failed | 0.0281 seconds  |
+./spec/models/lesson_spec.rb[1:1]                            | passed | 0.10868 seconds |
+./spec/models/lesson_spec.rb[1:2]                            | passed | 0.00624 seconds |
+./spec/models/lesson_spec.rb[1:3]                            | passed | 0.00584 seconds |
+./spec/models/lesson_spec.rb[1:4]                            | passed | 0.00657 seconds |
+./spec/models/lesson_spec.rb[1:5]                            | passed | 0.00335 seconds |
+./spec/models/lesson_spec.rb[1:6]                            | passed | 0.00636 seconds |
+./spec/models/lesson_spec.rb[1:7]                            | passed | 0.00986 seconds |
+./spec/models/lesson_term_spec.rb[1:1]                       | passed | 0.00321 seconds |
+./spec/models/lesson_term_spec.rb[1:2]                       | passed | 0.00652 seconds |
+./spec/models/term_spec.rb[1:1]                              | passed | 0.00442 seconds |
+./spec/models/term_spec.rb[1:2]                              | passed | 0.12046 seconds |
+./spec/models/term_spec.rb[1:3]                              | passed | 0.00414 seconds |
+./spec/models/term_spec.rb[1:4]                              | passed | 0.0025 seconds  |
