--- spec/views/terms/show.html.erb_spec.rb
@@ -1,16 +1,0 @@
-# frozen_string_literal: true
-
-require 'rails_helper'
-
-RSpec.describe 'terms/show', type: :view do
-  before(:each) do
-    @term = assign(:term, Term.create!(
-                            name: 'Name'
-    ))
-  end
-
-  it 'renders attributes in <p>' do
-    render
-    expect(rendered).to match(/Name/)
-  end
-end
